+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## \[0.1.20\] \[2020-20-25\]
+
+### Added
+
+-   Settings regarding picking grid added to `compas_rcf.fabrication.conf`.
+
+### Changed
+
+-   Function `get_picking_frame` in `compas_rcf.fabrication.abb_runner` reworked into `pick_frame_from_grid` which reads settings from conf to return frames from a grid of picking points.
+-   JSON dump after end of `abb_runner` now dumps original list (`clay_bullets`) instead of a new.
+
+## \[0.1.18\] \[2020-02-20\]
+
 ## \[0.1.19\] \[2020-20-22\]
 
 ### Changed